requirements.txt
fastapi
uvicorn
python-telegram-bot==13.7
schedule==1.1.0
# ...existing dependencies...